ShipmentScheduleUpdater: move into all relevant code from ShipmentScheduleBL; fix tests
https://github.com/metasfresh/metasfresh/issues/5679
