+++ de.metas.handlingunits.base/src/test/java/de/metas/handlingunits/impl/HUPickingSlotBL_RetrieveAvailableHUsToPickTests.java
@@ -37,9 +37,9 @@ import de.metas.handlingunits.picking.PickingCandidate;
 import de.metas.handlingunits.picking.PickingCandidateRepository;
 import de.metas.handlingunits.picking.PickingCandidateStatus;
 import de.metas.handlingunits.picking.impl.HUPickingSlotBL;
-import de.metas.inoutcandidate.api.IShipmentScheduleBL;
+import de.metas.inoutcandidate.api.IShipmentScheduleUpdater;
 import de.metas.inoutcandidate.api.ShipmentScheduleId;
-import de.metas.inoutcandidate.api.impl.ShipmentScheduleBL;
+import de.metas.inoutcandidate.api.impl.ShipmentScheduleUpdater;
 import de.metas.inoutcandidate.model.I_M_ShipmentSchedule;
 import de.metas.quantity.Quantity;
 import de.metas.storage.IStorageEngine;
@@ -95,7 +95,7 @@ public class HUPickingSlotBL_RetrieveAvailableHUsToPickTests
 
 		Services.get(IStorageEngineService.class).registerStorageEngine(storageEngine);
 
-		Services.registerService(IShipmentScheduleBL.class, ShipmentScheduleBL.newInstanceForUnitTesting());
+		Services.registerService(IShipmentScheduleUpdater.class, ShipmentScheduleUpdater.newInstanceForUnitTesting());
 
 		uom = newInstance(I_C_UOM.class);
 		saveRecord(uom);
