+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/inoutcandidate/api/impl/ShipmentScheduleUpdater.java
@@ -1,5 +1,11 @@
 package de.metas.inoutcandidate.api.impl;
 
+import static org.adempiere.model.InterfaceWrapperHelper.loadOutOfTrx;
+import static org.adempiere.model.InterfaceWrapperHelper.save;
+
+import java.math.BigDecimal;
+import java.time.ZonedDateTime;
+
 /*
  * #%L
  * de.metas.swat.base
@@ -13,28 +19,58 @@ package de.metas.inoutcandidate.api.impl;
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.Properties;
 import java.util.Set;
 
 import org.adempiere.ad.trx.api.ITrx;
+import org.adempiere.exceptions.AdempiereException;
+import org.adempiere.inout.util.DeliveryGroupCandidate;
+import org.adempiere.inout.util.DeliveryGroupCandidateGroupId;
+import org.adempiere.inout.util.DeliveryLineCandidate;
+import org.adempiere.inout.util.IShipmentSchedulesDuringUpdate;
+import org.adempiere.inout.util.IShipmentSchedulesDuringUpdate.CompleteStatus;
+import org.adempiere.inout.util.ShipmentScheduleAvailableStockDetail;
+import org.adempiere.inout.util.ShipmentScheduleQtyOnHandStorage;
+import org.adempiere.inout.util.ShipmentScheduleQtyOnHandStorageFactory;
+import org.adempiere.inout.util.ShipmentSchedulesDuringUpdate;
+import org.adempiere.model.InterfaceWrapperHelper;
+import org.adempiere.util.lang.IContextAware;
+import org.adempiere.warehouse.WarehouseId;
+import org.compiere.model.I_C_BPartner;
+import org.compiere.model.I_C_BPartner_Product;
+import org.compiere.util.TimeUtil;
 import org.slf4j.Logger;
+import org.springframework.stereotype.Service;
 
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 
+import de.metas.adempiere.model.I_M_Product;
+import de.metas.bpartner.BPartnerId;
+import de.metas.bpartner.BPartnerLocationId;
+import de.metas.bpartner.service.IBPartnerBL;
+import de.metas.bpartner_product.IBPartnerProductDAO;
+import de.metas.document.engine.DocStatus;
+import de.metas.inoutcandidate.api.IDeliverRequest;
+import de.metas.inoutcandidate.api.IShipmentConstraintsBL;
+import de.metas.inoutcandidate.api.IShipmentScheduleAllocBL;
+import de.metas.inoutcandidate.api.IShipmentScheduleAllocDAO;
 import de.metas.inoutcandidate.api.IShipmentScheduleBL;
+import de.metas.inoutcandidate.api.IShipmentScheduleEffectiveBL;
 import de.metas.inoutcandidate.api.IShipmentScheduleHandlerBL;
 import de.metas.inoutcandidate.api.IShipmentScheduleInvalidateRepository;
 import de.metas.inoutcandidate.api.IShipmentSchedulePA;
@@ -42,13 +78,54 @@ import de.metas.inoutcandidate.api.IShipmentScheduleUpdater;
 import de.metas.inoutcandidate.api.OlAndSched;
 import de.metas.inoutcandidate.api.ShipmentScheduleId;
 import de.metas.inoutcandidate.model.I_M_ShipmentSchedule;
+import de.metas.inoutcandidate.spi.IShipmentSchedulesAfterFirstPassUpdater;
+import de.metas.inoutcandidate.spi.ShipmentScheduleReferencedLine;
+import de.metas.inoutcandidate.spi.ShipmentScheduleReferencedLineFactory;
+import de.metas.inoutcandidate.spi.impl.CompositeCandidateProcessor;
+import de.metas.inoutcandidate.spi.impl.ShipmentScheduleOrderReferenceProvider;
+import de.metas.lang.SOTrx;
 import de.metas.logging.LogManager;
+import de.metas.material.cockpit.stock.StockRepository;
+import de.metas.order.DeliveryRule;
+import de.metas.organization.OrgId;
 import de.metas.process.PInstanceId;
+import de.metas.product.IProductBL;
+import de.metas.product.ProductId;
+import de.metas.quantity.Quantity;
+import de.metas.tourplanning.api.IDeliveryDayBL;
+import de.metas.tourplanning.api.IShipmentScheduleDeliveryDayBL;
+import de.metas.uom.IUOMConversionBL;
+import de.metas.uom.UomId;
 import de.metas.util.Check;
 import de.metas.util.Services;
+import lombok.NonNull;
 
+@Service
 public class ShipmentScheduleUpdater implements IShipmentScheduleUpdater
 {
+	@VisibleForTesting
+	public static ShipmentScheduleUpdater newInstanceForUnitTesting()
+	{
+		final StockRepository stockRepository = new StockRepository();
+		final ShipmentScheduleQtyOnHandStorageFactory shipmentScheduleQtyOnHandStorageFactory = new ShipmentScheduleQtyOnHandStorageFactory(stockRepository);
+		final ShipmentScheduleReferencedLineFactory shipmentScheduleReferencedLineFactory = new ShipmentScheduleReferencedLineFactory(Optional.of(ImmutableList.of(new ShipmentScheduleOrderReferenceProvider())));
+
+		return new ShipmentScheduleUpdater(
+				shipmentScheduleQtyOnHandStorageFactory,
+				shipmentScheduleReferencedLineFactory);
+	}
+
+	private static final String DYNATTR_ProcessedByBackgroundProcess = IShipmentScheduleBL.class.getName() + "#ProcessedByBackgroundProcess";
+
+	private static final Logger logger = LogManager.getLogger(ShipmentScheduleUpdater.class);
+	private final IShipmentScheduleInvalidateRepository invalidSchedulesRepo = Services.get(IShipmentScheduleInvalidateRepository.class);
+	private final IShipmentSchedulePA shipmentSchedulePA = Services.get(IShipmentSchedulePA.class);
+	private final IShipmentScheduleBL shipmentScheduleBL = Services.get(IShipmentScheduleBL.class);
+	private final ShipmentScheduleQtyOnHandStorageFactory shipmentScheduleQtyOnHandStorageFactory;
+	private final ShipmentScheduleReferencedLineFactory shipmentScheduleReferencedLineFactory;
+
+	private final CompositeCandidateProcessor candidateProcessors = new CompositeCandidateProcessor();
+
 	/**
 	 * Flag which is set to true when shipment schedule updater is running.
 	 *
@@ -56,13 +133,18 @@ public class ShipmentScheduleUpdater implements IShipmentScheduleUpdater
 	 */
 	private final InheritableThreadLocal<Boolean> running = new InheritableThreadLocal<>();
 
-	private static final Logger logger = LogManager.getLogger(ShipmentScheduleUpdater.class);
+	public ShipmentScheduleUpdater(
+			@NonNull final ShipmentScheduleQtyOnHandStorageFactory shipmentScheduleQtyOnHandStorageFactory,
+			@NonNull final ShipmentScheduleReferencedLineFactory shipmentScheduleReferencedLineFactory)
+	{
+		this.shipmentScheduleQtyOnHandStorageFactory = shipmentScheduleQtyOnHandStorageFactory;
+		this.shipmentScheduleReferencedLineFactory = shipmentScheduleReferencedLineFactory;
+	}
 
 	@Override
-	public int updateShipmentSchedule(final Properties ctx, final int adUserId, final PInstanceId adPInstanceId)
+	public void registerCandidateProcessor(final IShipmentSchedulesAfterFirstPassUpdater processor)
 	{
-		final boolean updateOnlyLocked = false;
-		return updateShipmentSchedule(ctx, adUserId, adPInstanceId, updateOnlyLocked);
+		candidateProcessors.addCandidateProcessor(processor);
 	}
 
 	@Override
@@ -73,9 +155,6 @@ public class ShipmentScheduleUpdater implements IShipmentScheduleUpdater
 			final boolean updateOnlyLocked)
 	{
 		// services
-		final IShipmentScheduleInvalidateRepository invalidSchedulesRepo = Services.get(IShipmentScheduleInvalidateRepository.class);
-		final IShipmentSchedulePA shipmentSchedulePA = Services.get(IShipmentSchedulePA.class);
-		final IShipmentScheduleBL shipmentScheduleBL = Services.get(IShipmentScheduleBL.class);
 
 		final Boolean running = this.running.get();
 		Check.assume(running == null || running == false, "updateShipmentSchedule is not already running");
@@ -97,7 +176,7 @@ public class ShipmentScheduleUpdater implements IShipmentScheduleUpdater
 			final List<OlAndSched> collectResult = retrieveOlsAndSchedsToProcess(adPInstanceId);
 
 			logger.debug("Invoking shipmentScheduleBL to update {} shipment schedule entries.", collectResult.size());
-			shipmentScheduleBL.updateSchedules(ctx, collectResult, ITrx.TRXNAME_ThreadInherited);
+			updateSchedules(ctx, collectResult);
 
 			// cleanup the marker/pointer tables
 			invalidSchedulesRepo.deleteRecomputeMarkersOutOfTrx(adPInstanceId);
@@ -145,4 +224,675 @@ public class ShipmentScheduleUpdater implements IShipmentScheduleUpdater
 		logger.debug("Found additional {} schedule entries to update", (collectResult.size() - olsAndScheds.size()));
 		return collectResult;
 	}
+
+	/**
+	 * Updates the given {@link I_M_ShipmentSchedule}s by setting these columns:
+	 * <li>
+	 * {@link I_M_ShipmentSchedule#COLUMNNAME_QtyToDeliver}
+	 * <li>
+	 * {@link I_M_ShipmentSchedule#COLUMNNAME_QtyOnHand}
+	 * <li>
+	 * {@link I_M_ShipmentSchedule#COLUMNNAME_Status}
+	 * <li>
+	 * {@link I_M_ShipmentSchedule#COLUMNNAME_PostageFreeAmt}
+	 * <li>
+	 * {@link I_M_ShipmentSchedule#COLUMNNAME_AllowConsolidateInOut}
+	 *
+	 * To actually set those values, this method calls the registered {@link IShipmentSchedulesAfterFirstPassUpdater}.
+	 *
+	 *
+	 * @param olsAndScheds
+	 */
+	@VisibleForTesting
+	void updateSchedules(final Properties ctx, final List<OlAndSched> olsAndScheds)
+	{
+		if (olsAndScheds.isEmpty())
+		{
+			return;
+		}
+
+		//
+		// Services
+		final IBPartnerBL bpartnerBL = Services.get(IBPartnerBL.class);
+		final IShipmentScheduleDeliveryDayBL shipmentScheduleDeliveryDayBL = Services.get(IShipmentScheduleDeliveryDayBL.class);
+		final IShipmentScheduleEffectiveBL shipmentScheduleEffectiveBL = Services.get(IShipmentScheduleEffectiveBL.class);
+
+		//
+		// Briefly update our shipment schedules:
+		// * set BPartnerAddress_Override if was not set before
+		// * update HeaderAggregationKey
+		for (final OlAndSched olAndSched : olsAndScheds)
+		{
+			final I_M_ShipmentSchedule sched = olAndSched.getSched();
+
+			updateCatchUomId(sched);
+
+			updateWarehouseId(sched);
+
+			shipmentScheduleBL.updateBPArtnerAddressOverrideIfNotYetSet(sched);
+
+			shipmentScheduleBL.updateHeaderAggregationKey(sched);
+
+			updateShipmentConstraints(sched);
+		}
+
+		final ShipmentSchedulesDuringUpdate firstRun = generate(ctx, olsAndScheds, null);
+		firstRun.updateCompleteStatusAndSetQtyToZeroWhereNeeded();
+
+		final int removeCnt = applyCandidateProcessors(ctx, firstRun);
+		logger.info("{} records were discarded by candidate processors", removeCnt);
+
+		// evaluate the processor's result: lines that have been discarded won't
+		// be delivered and won't be validated in the second run.
+		for (final DeliveryLineCandidate inOutLine : firstRun.getAllLines())
+		{
+			if (inOutLine.isDiscarded())
+			{
+				inOutLine.setQtyToDeliver(BigDecimal.ZERO);
+			}
+			else
+			{
+				// remember: 'removeLine' means that a *new* line might be
+				// created for the corresponding olAndSched
+				inOutLine.removeFromGroup();
+				firstRun.removeLine(inOutLine);
+			}
+		}
+
+		// make the second run
+		final IShipmentSchedulesDuringUpdate secondRun = generate(ctx, olsAndScheds, firstRun);
+
+		// finally update the shipment schedule entries
+		for (final OlAndSched olAndSched : olsAndScheds)
+		{
+			final I_M_ShipmentSchedule sched = olAndSched.getSched();
+			final IDeliverRequest deliverRequest = olAndSched.getDeliverRequest();
+			final I_C_BPartner bPartner = shipmentScheduleEffectiveBL.getBPartner(sched); // task 08756: we don't really care for the ol's partner, but for the partner who will actually receive the shipment.
+
+			final boolean isBPAllowConsolidateInOut = bpartnerBL.isAllowConsolidateInOutEffective(bPartner, SOTrx.SALES);
+			sched.setAllowConsolidateInOut(isBPAllowConsolidateInOut);
+
+			updatePreparationAndDeliveryDate(sched);
+
+			//
+			// Delivery Day related info:
+			// TODO: invert dependency add make this pluggable from de.metas.tourplanning module
+			shipmentScheduleDeliveryDayBL.updateDeliveryDayInfo(sched);
+
+			// task 09358: ol.qtyReserved should be as correct as QtyOrdered and QtyDelivered, but in some cases isn't. this here is a workaround to the problem
+			// task 09869: don't rely on ol anyways
+			final BigDecimal qtyDelivered = Services.get(IShipmentScheduleAllocDAO.class).retrieveQtyDelivered(sched);
+			sched.setQtyDelivered(qtyDelivered);
+			sched.setQtyReserved(BigDecimal.ZERO.max(deliverRequest.getQtyOrdered().subtract(sched.getQtyDelivered())));
+
+			updateLineNetAmt(olAndSched);
+
+			ShipmentScheduleQtysHelper.updateQtyToDeliver(olAndSched, secondRun);
+
+			markAsChangedByUpdateProcess(sched);
+
+			if (olAndSched.hasSalesOrderLine())
+			{
+				final DocStatus orderDocStatus = olAndSched.getOrderDocStatus();
+				if (!orderDocStatus.isCompletedOrClosedOrReversed() // task 07355: thread closed orders like completed orders
+						&& !sched.isProcessed() // task 05206: ts: don't try to delete already processed scheds..it won't work
+						&& sched.getQtyDelivered().signum() == 0 // also don't try to delete if there is already a picked or delivered Qty.
+						&& sched.getQtyPickList().signum() == 0)
+				{
+					logger.debug("QtyToDeliver_Override=" + sched.getQtyToDeliver_Override()
+							+ "; QtyReserved=" + sched.getQtyReserved()
+							+ "; DocStatus=" + orderDocStatus
+							+ "; => Deleting " + sched);
+
+					InterfaceWrapperHelper.delete(sched);
+					continue;
+				}
+			}
+
+			updateProcessedFlag(sched);
+			if (sched.isProcessed())
+			{
+				// 04870 : Delivery rule force assumes we deliver full quantity ordered if qtyToDeliver_Override is null.
+				// 06019 : check both DeliveryRule, as DeliveryRule_Override
+				final boolean deliveryRuleIsForced = DeliveryRule.FORCE.equals(shipmentScheduleEffectiveBL.getDeliveryRule(sched));
+				if (deliveryRuleIsForced)
+				{
+					sched.setQtyToDeliver(BigDecimal.ZERO);
+					save(sched);
+				}
+				else
+				{
+					Check.errorUnless(sched.getQtyToDeliver().signum() == 0, "{} has QtyToDeliver = {} (should be zero)", sched, sched.getQtyToDeliver());
+				}
+				continue;
+			}
+
+			// task 08694
+			// I talked with Mark and he observed that in the wiki-page of 08459 it is specified differently.
+			// I will let it here nevertheless, so we can keep track of it's way to work
+
+			final BPartnerId partnerId = BPartnerId.ofRepoId(sched.getC_BPartner_ID());
+
+			// FRESH-334 retrieve the bp product for org or for org 0
+			final org.compiere.model.I_M_Product product = loadOutOfTrx(sched.getM_Product_ID(), I_M_Product.class);
+			final OrgId orgId = OrgId.ofRepoId(product.getAD_Org_ID());
+
+			final I_C_BPartner_Product bpp = Services.get(IBPartnerProductDAO.class).retrieveBPartnerProductAssociation(ctx, partnerId, ProductId.ofRepoId(sched.getM_Product_ID()), orgId);
+			if (bpp == null)
+			{
+				// in case no dropship bpp entry was found, the schedule shall not be dropship
+				sched.setIsDropShip(false);
+			}
+			else
+			{
+
+				final boolean isDropShip = bpp.isDropShip();
+
+				if (isDropShip)
+				{
+					// if there is bpp that is dropship and has a C_BPartner_Vendor_ID,
+					// set the customer's vendor for the given product in the schedule
+					sched.setC_BPartner_Vendor_ID(bpp.getC_BPartner_Vendor_ID());
+				}
+
+				// set the dropship flag in shipment schedule as it is in the bpp
+				sched.setIsDropShip(isDropShip);
+			}
+
+			// 08860
+			// update preparation date override based on delivery date effective
+			// DO this only if the preparationDate_Override was not already set manually or by the process
+
+			if (sched.getDeliveryDate_Override() != null && sched.getPreparationDate_Override() == null)
+			{
+				final ZonedDateTime deliveryDate = shipmentScheduleEffectiveBL.getDeliveryDate(sched);
+
+				final IDeliveryDayBL deliveryDayBL = Services.get(IDeliveryDayBL.class);
+				final IContextAware contextAwareSched = InterfaceWrapperHelper.getContextAware(sched);
+				final BPartnerLocationId bpLocationId = shipmentScheduleEffectiveBL.getBPartnerLocationId(sched);
+
+				final ZonedDateTime calculationTime = TimeUtil.asZonedDateTime(sched.getCreated());
+				final ZonedDateTime preparationDate = deliveryDayBL.calculatePreparationDateOrNull(
+						contextAwareSched,
+						SOTrx.SALES,
+						calculationTime,
+						deliveryDate,
+						bpLocationId);
+
+				// In case the DeliveryDate Override is set, also update the preparationDate override
+				sched.setPreparationDate_Override(TimeUtil.asTimestamp(preparationDate));
+
+			}
+			save(sched);
+		}
+	}
+
+	ShipmentSchedulesDuringUpdate generate(
+			final Properties ctx,
+			final List<OlAndSched> lines,
+			final ShipmentSchedulesDuringUpdate firstRun)
+	{
+		// services
+		final IShipmentScheduleEffectiveBL shipmentScheduleEffectiveValuesBL = Services.get(IShipmentScheduleEffectiveBL.class);
+		final IShipmentScheduleAllocBL shipmentScheduleAllocBL = Services.get(IShipmentScheduleAllocBL.class);
+		final IProductBL productBL = Services.get(IProductBL.class);
+
+		// if firstRun is not null, create a new instance, otherwise use firstRun
+		final ShipmentSchedulesDuringUpdate candidates = mkCandidatesToUse(lines, firstRun);
+
+		//
+		// Load QtyOnHand in scope for our lines
+		// i.e. iterate all lines to cache the required storage info and to subtract the quantities that can't be allocated from the storage allocation.
+		final ShipmentScheduleQtyOnHandStorage qtyOnHands = shipmentScheduleQtyOnHandStorageFactory.ofOlAndScheds(lines);
+
+		//
+		// Iterate again and:
+		// * try to allocate the QtyOnHand
+		for (final OlAndSched olAndSched : lines)
+		{
+			final I_M_ShipmentSchedule sched = olAndSched.getSched();
+			final IDeliverRequest deliverRequest = olAndSched.getDeliverRequest();
+
+			final DeliveryRule deliveryRule = shipmentScheduleEffectiveValuesBL.getDeliveryRule(sched);
+
+			final boolean ruleManual = DeliveryRule.MANUAL.equals(deliveryRule);
+
+			final BigDecimal qtyRequired;
+			if (olAndSched.getQtyOverride() != null)
+			{
+				qtyRequired = olAndSched.getQtyOverride().subtract(ShipmentScheduleQtysHelper.mkQtyToDeliverOverrideFulFilled(olAndSched));
+			}
+			else if (ruleManual)
+			{
+				// lines with ruleManual need to be scheduled explicitly using QtyToDeliver_Override
+				logger.debug("Manual - QtyOverride not set");
+				qtyRequired = BigDecimal.ZERO;
+			}
+			else
+			{
+				final BigDecimal qtyDelivered = Services.get(IShipmentScheduleAllocDAO.class).retrieveQtyDelivered(sched);
+				qtyRequired = deliverRequest.getQtyOrdered().subtract(qtyDelivered);
+			}
+
+			//
+			// QtyPickList (i.e. qtyUnconfirmedShipments) is the sum of
+			// * MovementQtys from all draft shipment lines which are pointing to shipment schedule's order line
+			// * QtyPicked from QtyPicked records
+			final ProductId productId = ProductId.ofRepoId(sched.getM_Product_ID());
+
+			final BigDecimal qtyPickList;
+			{
+				// task 08123: we also take those numbers into account that are *not* on an M_InOutLine yet, but are nonetheless picked
+				final Quantity stockingQty = shipmentScheduleAllocBL.retrieveQtyPickedAndUnconfirmed(sched);
+				qtyPickList = stockingQty.toBigDecimal();
+
+				// Update shipment schedule's fields
+				sched.setQtyPickList(qtyPickList);
+			}
+
+			final BigDecimal qtyToDeliver = ShipmentScheduleQtysHelper.mkQtyToDeliver(qtyRequired, qtyPickList);
+
+			if (!productBL.isStocked(productId))
+			{
+				// product not stocked => don't concern ourselves with the storage; just deliver what was ordered
+				createLine(ctx, olAndSched, qtyToDeliver,
+						null/* storages */,
+						true/* force */,
+						CompleteStatus.OK,
+						candidates);
+				continue;
+			}
+
+			//
+			// Get the QtyOnHand storages suitable for our order line
+			final List<ShipmentScheduleAvailableStockDetail> storages = qtyOnHands.getStockDetailsMatching(sched);
+			final BigDecimal qtyOnHandBeforeAllocation = ShipmentScheduleAvailableStockDetail.calculateQtyOnHandSum(storages);
+			sched.setQtyOnHand(qtyOnHandBeforeAllocation);
+
+			final CompleteStatus completeStatus = mkCompleteStatus(qtyToDeliver, qtyOnHandBeforeAllocation);
+			final boolean ruleCompleteOrder = DeliveryRule.COMPLETE_ORDER.equals(deliveryRule);
+			final boolean ruleAvailable = DeliveryRule.AVAILABILITY.equals(deliveryRule);
+			final boolean ruleCompleteLine = DeliveryRule.COMPLETE_LINE.equals(deliveryRule);
+			final boolean ruleForce = DeliveryRule.FORCE.equals(deliveryRule);
+
+			if (ruleForce)
+			{
+				// delivery rule force
+				logger.debug("Force - OnHand=" + qtyOnHandBeforeAllocation
+						+ " (QtyPickList=" + qtyPickList + "), ToDeliver=" + qtyToDeliver
+						+ ", Delivering=" + qtyToDeliver);
+
+				createLine(ctx, olAndSched, qtyToDeliver, storages, ruleForce, completeStatus, candidates);
+			}
+			// delivery rules "Complete" and "Availability"
+			else if (ruleCompleteOrder || ruleCompleteLine || ruleAvailable || ruleManual)
+			{
+				if (qtyOnHandBeforeAllocation.signum() > 0 || qtyToDeliver.signum() < 0)
+				{
+					//
+					// if there is anything at all to deliver, create a new
+					// inOutLine
+					BigDecimal deliver = qtyToDeliver;
+					if (deliver.compareTo(qtyOnHandBeforeAllocation) > 0)
+					{
+						deliver = qtyOnHandBeforeAllocation;
+					}
+
+					// we invoke createLine even if ruleComplete is true and
+					// fullLine is false, because we want the quantity to be
+					// allocated.
+					// If the created line will make it into a real shipment will be
+					// decided later.
+					logger.debug("CompleteLine - OnHand=" + qtyOnHandBeforeAllocation
+							+ " (QtyPickList=" + qtyPickList + "), ToDeliver=" + qtyToDeliver
+							+ ", FullLine=" + completeStatus);
+
+					createLine(ctx, olAndSched, deliver, storages, ruleForce, completeStatus, candidates);
+				}
+				else
+				{
+					logger.debug("No qtyOnHand to deliver[SKIP] - OnHand=" + qtyOnHandBeforeAllocation
+							+ " (QtyPickList=" + qtyPickList + "), ToDeliver=" + qtyToDeliver
+							+ ", FullLine=" + completeStatus);
+				}
+			}
+			else
+			{
+				throw new AdempiereException(
+						"Unsupported delivery rule: " + deliveryRule
+								+ " - OnHand=" + qtyOnHandBeforeAllocation + " (PickedNotDelivered=" + qtyPickList + "), ToDeliver=" + qtyToDeliver);
+			}
+		}
+
+		logger.info("Created " + candidates.size() + " shipment candidates");
+
+		return candidates;
+	} // generate
+
+	private CompleteStatus mkCompleteStatus(final BigDecimal toDeliver, final BigDecimal onHand)
+	{
+		CompleteStatus completeStatus = CompleteStatus.OK;
+		{
+			// fullLine==true means that we can create a shipment line with
+			// the full (current!) quantity of the current order line
+			final boolean fullLine = onHand.compareTo(toDeliver) >= 0 || toDeliver.signum() <= 0;
+			if (!fullLine)
+			{
+				completeStatus = CompleteStatus.INCOMPLETE_LINE;
+			}
+		}
+		return completeStatus;
+	}
+
+	private ShipmentSchedulesDuringUpdate mkCandidatesToUse(
+			final List<OlAndSched> lines,
+			final ShipmentSchedulesDuringUpdate firstRun)
+	{
+		if (firstRun != null)
+		{
+			return firstRun;
+		}
+
+		return new ShipmentSchedulesDuringUpdate();
+
+	}
+
+	/**
+	 * Creates one or more inOutLines for a given orderLine.
+	 *
+	 * @param qty the quantity all created inOutLines' qtyEntered will sum up to
+	 */
+	private void createLine(
+			final Properties ctx,
+			final OlAndSched olAndSched,
+			final BigDecimal qty,
+			final List<ShipmentScheduleAvailableStockDetail> storages,
+			final boolean force,
+			final CompleteStatus completeStatus,
+			final ShipmentSchedulesDuringUpdate candidates)
+	{
+		final I_M_ShipmentSchedule sched = olAndSched.getSched();
+
+		if (candidates.getInOutLineFor(sched) != null)
+		{
+			logger.debug("candidates contains already an inoutLine for M_ShipmentSchedule {}", sched);
+			return;
+		}
+
+		final DeliveryGroupCandidate groupCandidate = getOrCreateGroupCandidateForShipmentSchedule(sched, candidates);
+
+		if (storages == null || storages.isEmpty())
+		{
+			final DeliveryLineCandidate deliveryLineCandidate = groupCandidate.createAndAddLineCandidate(sched, completeStatus);
+			if (force) // Case: no Quantity on Hand storages and force => no need for allocations etc
+			{
+				deliveryLineCandidate.setQtyToDeliver(qty);
+			}
+			candidates.addLine(deliveryLineCandidate);
+			return;
+		}
+
+		// Shipment Lines (i.e. candidate lines)
+		final List<DeliveryLineCandidate> inoutLines = new ArrayList<>();
+
+		//
+		// Iterate QtyOnHand storage records and try to allocate on current shipment schedule/order line.
+		BigDecimal toDeliverTargetQty = qty; // how much still needs to be delivered; initially it's the whole qty required
+		for (int i = 0; i < storages.size(); i++)
+		{
+			// Stop here is there is nothing remaining to be delivered
+			if (toDeliverTargetQty.signum() == 0)
+			{
+				break;
+			}
+
+			final ShipmentScheduleAvailableStockDetail storage = storages.get(i);
+			BigDecimal toDeliverCurrentQty = toDeliverTargetQty; // initially try to deliver the entire quantity remaining to be delivered
+
+			//
+			// Adjust the quantity that can be delivered from this storage line
+			// Check: Not enough On Hand
+			final BigDecimal qtyOnHandAvailable = storage.getQtyOnHand();
+			if (toDeliverCurrentQty.compareTo(qtyOnHandAvailable) > 0
+					&& qtyOnHandAvailable.signum() >= 0)         // positive storage
+			{
+				if (!force // Adjust to OnHand Qty
+						|| force && i + 1 != storages.size())         // if force not on last location
+				{
+					toDeliverCurrentQty = qtyOnHandAvailable;
+				}
+			}
+			// Skip if we cannot deliver something for current storage line
+			if (toDeliverCurrentQty.signum() == 0)
+			{
+				// zero deliver
+				continue;
+			}
+
+			// Find existing lineCandidate that was added in a previous iteration
+			DeliveryLineCandidate lineCandidate = null;
+			for (final DeliveryLineCandidate existingLineCandidate : inoutLines)
+			{
+				// skip if it's for a different order line
+				if (existingLineCandidate.getShipmentScheduleId() != sched.getM_ShipmentSchedule_ID())
+				{
+					continue;
+				}
+				lineCandidate = existingLineCandidate;
+			}
+
+			// Case: No InOutLine found
+			// => create a new InOutLine
+			if (lineCandidate == null)
+			{
+				lineCandidate = groupCandidate.createAndAddLineCandidate(sched, completeStatus);
+
+				lineCandidate.setQtyToDeliver(toDeliverCurrentQty);
+
+				inoutLines.add(lineCandidate);
+				candidates.addLine(lineCandidate);
+			}
+			//
+			// Case: existing InOutLine found
+			// => adjust the quantity
+			else
+			{
+				final BigDecimal inoutLineQtyOld = lineCandidate.getQtyToDeliver();
+				final BigDecimal inoutLineQtyNew = inoutLineQtyOld.add(toDeliverCurrentQty);
+				lineCandidate.setQtyToDeliver(inoutLineQtyNew);
+			}
+
+			logger.debug("ToDeliver=" + qty + "/" + toDeliverCurrentQty + " - " + lineCandidate);
+			toDeliverTargetQty = toDeliverTargetQty.subtract(toDeliverCurrentQty);
+
+			storage.subtractQtyOnHand(toDeliverCurrentQty);
+		}    // for each storage record
+	}
+
+	private DeliveryGroupCandidate getOrCreateGroupCandidateForShipmentSchedule(
+			@NonNull final I_M_ShipmentSchedule sched,
+			final IShipmentSchedulesDuringUpdate candidates)
+	{
+		final IBPartnerBL bpartnerBL = Services.get(IBPartnerBL.class);
+		final I_C_BPartner partner = loadOutOfTrx(sched.getC_BPartner_ID(), I_C_BPartner.class);
+
+		final ShipmentScheduleReferencedLine scheduleSourceDoc = shipmentScheduleReferencedLineFactory.createFor(sched);
+		final String bpartnerAddress = sched.getBPartnerAddress_Override();
+
+		DeliveryGroupCandidate candidate = null;
+
+		final WarehouseId warehouseId = Services.get(IShipmentScheduleEffectiveBL.class).getWarehouseId(sched);
+		final boolean consolidateAllowed = bpartnerBL.isAllowConsolidateInOutEffective(partner, SOTrx.SALES);
+		if (consolidateAllowed)
+		{
+			// see if there is an existing shipment for this location and shipper
+			candidate = candidates.getInOutForShipper(scheduleSourceDoc.getShipperId(), warehouseId, bpartnerAddress);
+		}
+		else
+		{
+			// see if there is an existing shipment for this order (or flatrate term)
+			candidate = candidates.getInOutForRecordRef(scheduleSourceDoc.getRecordRef(), warehouseId, bpartnerAddress);
+		}
+
+		if (candidate == null)
+		{
+			// create a new Shipment
+			candidate = createGroup(scheduleSourceDoc, sched);
+			candidates.addGroup(candidate);
+		}
+		return candidate;
+	}
+
+	@VisibleForTesting
+	DeliveryGroupCandidate createGroup(
+			final ShipmentScheduleReferencedLine scheduleSourceDoc,
+			final I_M_ShipmentSchedule sched)
+	{
+
+		final IShipmentScheduleEffectiveBL shipmentScheduleEffectiveBL = Services.get(IShipmentScheduleEffectiveBL.class);
+
+		return DeliveryGroupCandidate.builder()
+				.warehouseId(shipmentScheduleEffectiveBL.getWarehouseId(sched))
+				.bPartnerAddress(sched.getBPartnerAddress_Override())
+				.groupId(DeliveryGroupCandidateGroupId.of(scheduleSourceDoc.getRecordRef()))
+				.shipperId(scheduleSourceDoc.getShipperId())
+				.build();
+	}
+
+	/**
+	 *
+	 * @param sched
+	 * @return
+	 * @task 08336
+	 */
+	@VisibleForTesting
+	void updateProcessedFlag(@NonNull final I_M_ShipmentSchedule sched)
+	{
+		if (sched.isClosed())
+		{
+			sched.setProcessed(true);
+			return;
+
+		}
+		final boolean noQtyOverride = sched.getQtyToDeliver_Override().signum() <= 0;
+		final boolean noQtyReserved = sched.getQtyReserved().signum() <= 0;
+
+		sched.setProcessed(noQtyOverride && noQtyReserved);
+	}
+
+	private void updatePreparationAndDeliveryDate(@NonNull final I_M_ShipmentSchedule sched)
+	{
+		final ShipmentScheduleReferencedLine shipmentScheduleOrderDoc = shipmentScheduleReferencedLineFactory.createFor(sched);
+
+		sched.setPreparationDate(TimeUtil.asTimestamp(shipmentScheduleOrderDoc.getPreparationDate()));
+		sched.setDeliveryDate(TimeUtil.asTimestamp(shipmentScheduleOrderDoc.getDeliveryDate()));
+	}
+
+	private void updateShipmentConstraints(final I_M_ShipmentSchedule sched)
+	{
+		final IShipmentConstraintsBL shipmentConstraintsBL = Services.get(IShipmentConstraintsBL.class);
+		final int billBPartnerId = sched.getBill_BPartner_ID();
+		final int deliveryStopShipmentConstraintId = shipmentConstraintsBL.getDeliveryStopShipmentConstraintId(billBPartnerId);
+		final boolean isDeliveryStop = deliveryStopShipmentConstraintId > 0;
+		if (isDeliveryStop)
+		{
+			sched.setIsDeliveryStop(true);
+			sched.setM_Shipment_Constraint_ID(deliveryStopShipmentConstraintId);
+		}
+		else
+		{
+			sched.setIsDeliveryStop(false);
+			sched.setM_Shipment_Constraint_ID(-1);
+		}
+	}
+
+	private void updateLineNetAmt(final OlAndSched olAndSched)
+	{
+		final BigDecimal lineNetAmt = computeLineNetAmt(olAndSched);
+		olAndSched.setShipmentScheduleLineNetAmt(lineNetAmt);
+	}
+
+	/**
+	 * Try to get the given <code>ol</code>'s <code>qtyReservedInPriceUOM</code> and update the given <code>sched</code>'s <code>LineNetAmt</code>.
+	 *
+	 * @task https://github.com/metasfresh/metasfresh/issues/298
+	 * @throws AdempiereException in developer mode, if there the <code>qtyReservedInPriceUOM</code> can't be obtained.
+	 */
+	private BigDecimal computeLineNetAmt(final OlAndSched olAndSched)
+	{
+		if (!olAndSched.hasSalesOrderLine())
+		{
+			return BigDecimal.ZERO;
+		}
+
+		final IUOMConversionBL uomConversionBL = Services.get(IUOMConversionBL.class);
+		// final de.metas.interfaces.I_C_OrderLine olEx = InterfaceWrapperHelper.create(ol, de.metas.interfaces.I_C_OrderLine.class);
+		final BigDecimal qtyReservedInPriceUOM = uomConversionBL.convertFromProductUOM(
+				olAndSched.getProductId(),
+				olAndSched.getOrderPriceUOM(),
+				olAndSched.getOrderQtyReserved());
+
+		// qtyReservedInPriceUOM might be null. in that case, don't fail the whole updating, but set the value to null
+		if (qtyReservedInPriceUOM == null)
+		{
+			final String msg = "IUOMConversionBL.convertFromProductUOM() failed for " + olAndSched + "; \n"
+					+ "Therefore we can't set LineNetAmt for M_ShipmentSchedule; \n"
+					+ "Note: if this exception was thrown and not just logged, then check for stale M_ShipmentSchedule_Recompute records";
+			new AdempiereException(msg).throwIfDeveloperModeOrLogWarningElse(logger);
+			return BigDecimal.ONE.negate();
+		}
+		else
+		{
+			final BigDecimal orderPriceActual = olAndSched.getOrderPriceActual();
+			return qtyReservedInPriceUOM.multiply(orderPriceActual);
+		}
+	}
+
+	private int applyCandidateProcessors(final Properties ctx, final IShipmentSchedulesDuringUpdate candidates)
+	{
+		return candidateProcessors.doUpdateAfterFirstPass(ctx, candidates, ITrx.TRXNAME_ThreadInherited);
+	}
+
+	/**
+	 * 07400 also update the M_Warehouse_ID; an order might have been reactivated and the warehouse might have been changed.
+	 *
+	 * @param sched
+	 */
+	private void updateWarehouseId(@NonNull final I_M_ShipmentSchedule sched)
+	{
+		final WarehouseId warehouseId = shipmentScheduleReferencedLineFactory
+				.createFor(sched)
+				.getWarehouseId();
+		sched.setM_Warehouse_ID(warehouseId.getRepoId());
+	}
+
+	private void updateCatchUomId(@NonNull final I_M_ShipmentSchedule sched)
+	{
+		final IProductBL productBL = Services.get(IProductBL.class);
+
+		final boolean isCatchWeight = shipmentScheduleBL.isCatchWeight(sched);
+		if (!isCatchWeight)
+		{
+			sched.setCatch_UOM_ID(-1);
+			return;
+		}
+
+		final Optional<UomId> catchUOMId = productBL.getCatchUOMId(ProductId.ofRepoId(sched.getM_Product_ID()));
+		final Integer catchUomRepoId = catchUOMId.map(UomId::getRepoId).orElse(0);
+
+		sched.setCatch_UOM_ID(catchUomRepoId);
+	}
+
+	@Override
+	public boolean isChangedByUpdateProcess(final I_M_ShipmentSchedule sched)
+	{
+		final Boolean isUpdateProcess = InterfaceWrapperHelper.getDynAttribute(sched, DYNATTR_ProcessedByBackgroundProcess);
+		return isUpdateProcess == null ? false : isUpdateProcess.booleanValue();
+	}
+
+	private void markAsChangedByUpdateProcess(final I_M_ShipmentSchedule sched)
+	{
+		InterfaceWrapperHelper.setDynAttribute(sched, DYNATTR_ProcessedByBackgroundProcess, Boolean.TRUE);
+	}
+
 }
