+++ de.metas.swat/de.metas.swat.base/src/test/java/de/metas/tourplanning/integrationtest/TourInstance_DeliveryDay_ShipmentSchedule_IntegrationTest.java
@@ -32,7 +32,7 @@ import org.adempiere.model.InterfaceWrapperHelper;
 import org.compiere.model.I_C_BPartner;
 import org.compiere.model.I_C_BPartner_Location;
 import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import de.metas.adempiere.model.I_C_Order;
 import de.metas.adempiere.model.I_M_Product;
