+++ de.metas.swat/de.metas.swat.base/src/test/java/de/metas/tourplanning/integrationtest/TourPlanning_PreparationDate_IntergrationTest.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import org.compiere.model.I_C_BPartner;
 import org.compiere.model.I_C_BPartner_Location;
 import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import de.metas.adempiere.model.I_C_Order;
 import de.metas.tourplanning.TourPlanningTestBase;
